* adds deducated folders for list and menu * moves files
