+++ src/app/users/routing.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { UsersComponent } from './container/component/component';
-import { UsersMenuComponent } from './container/menu/component';
-import { UsersWithPresenterComponent } from './container/presenter/component';
+import { UsersComponent } from './list/component/component';
+import { UsersWithPresenterComponent } from './list/presenter/component';
+import { UsersMenuComponent } from './menu/component';
 
 const routes: Routes = [
   {
